Refactoring and minor changes * fixed signatures of ReadCPACS/WriteCPACS * added CCPACSFramesAssembly::Invalidate * bounded cut shape of CCPACSFrame * removed max macro * removed unneeded whitespace * removed unused variables * added CCPACSGuideCurve::GetFromDefinition() * added TODOs * added CCPACSPressureBulkhead::GetReinforcementNumberVertical/Horizontal with defaulted values * replaced explicit optional<T> access by calling value() * removed deprecated CCPACSWingComponentSegment::GetMidplaneEtaXsi * added CCPACSWingComponentSegment::GetEtaXsiLocal for local points * removed a method declaration with no definition

+++ src/common/tiglcommonfunctions.cpp
@@ -497,7 +497,7 @@ int GetComponentHashCode(tigl::ITiglGeometricComponent& component)
 // Creates an Edge from the given Points by B-Spline interpolation
 TopoDS_Edge EdgeSplineFromPoints(const std::vector<gp_Pnt>& points)
 {
-    unsigned int pointCount = points.size();
+    unsigned int pointCount = static_cast<int>(points.size());
     
     Handle(TColgp_HArray1OfPnt) hpoints = new TColgp_HArray1OfPnt(1, pointCount);
     for (unsigned int j = 0; j < pointCount; j++) {
